Incorporated 2015 FTC 8863 team code into V2.3 ftc app using new Teamcode directory structure
Changed import and package references so they resolved correctly.
Changed enums that were relocated in V2.3 ftc app so they resolved to the correct class in the ftc app.
Restructured directories in teamcode so they were more intuitive.
Although the code builds, IT HAS NOT BEEN TESTED. It may not be working.